Change the action obtained when token is at goal
Beforehand, a -1 was written for avoiding that token to be moved. Now,
the correct value (4) is written, but a condition on the top-layer code
prevents it to be used.

The R matrix is stored in a text file, and it is imported, instead of
being stored in the code itself. The function for reading and writing
the Q and R matrices are made generic.

The discount_factor is now public, and a test is prepared for assessing
its effect.